Introduce new module for downloading custom configs + refactor downloadCustomConfig service #2734
