+++ visualization/app/codeCharta/ui/customConfigs/downloadCustomConfigsButton/downloadCustomConfigButton.module.ts
@@ -1,0 +1,12 @@
+import { NgModule } from "@angular/core"
+
+import { DownloadCustomConfigService } from "./downloadCustomConfig.service"
+import { DownloadCustomConfigsButtonComponent } from "./downloadCustomConfigsButton.component"
+
+@NgModule({
+	declarations: [DownloadCustomConfigsButtonComponent],
+	exports: [DownloadCustomConfigsButtonComponent],
+	providers: [DownloadCustomConfigService],
+	entryComponents: [DownloadCustomConfigsButtonComponent]
+})
+export class DownloadCustomConfigButtonModule {}
